put admin check on delete play/skip

+++ src/Application/Api/Ajax/Handler/IndexAjaxHandler.php
@@ -337,7 +337,9 @@ final readonly class IndexAjaxHandler implements AjaxHandlerInterface
 
                 break;
             case 'delete_play':
-                if (isset($_REQUEST['activity_id'])) {
+                if (Access::check(AccessTypeEnum::INTERFACE, AccessLevelEnum::ADMIN) &&
+                    isset($_REQUEST['activity_id'])
+                ) {
                     Stats::delete((int)$_REQUEST['activity_id']);
                 }
